Small fixes and created migrations for the database
Me adding in migrations does not mean your data is save in the database.

+++ JustDanceNextPlus/JustDanceClasses/Database/Profile/MapStats.cs
@@ -14,7 +14,7 @@ public class MapStats
 	public Guid ProfileId { get; set; }
 
 	public int HighScore { get; set; }
-	public long PlayCount { get; set; }
+	public int PlayCount { get; set; }
 	public string Platform { get; set; } = "Unknown";
 	public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
